pyproject.toml
[project]
name = "agentic_investment_advisor"
version = "0.1.0"
description = "agentic_investment_advisor using crewAI"
authors = [{ name = "Sebastian Cardona", email = "sebas.c@hotmail.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.2.1",
    "litellm>=1.79.0",
    "opik>=1.8.68",
    "portkey-ai>=2.0.2",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.12",
]

[project.scripts]
agentic_investment_advisor = "agentic_investment_advisor.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pip>=25.3",
    "ruff>=0.14.3",
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
